# src/openApi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "Caveman Chartplotter API",
    "version": "1.0.0",
    "description": "HTTP API for the Caveman Chartplotter plugin: UI configuration and the custom own-boat icon.",
    "license": {
      "name": "Apache-2.0",
      "url": "https://www.apache.org/licenses/LICENSE-2.0"
    }
  },
  "servers": [
    {
      "url": "/plugins/caveman-chartplotter"
    }
  ],
  "tags": [
    {
      "name": "config",
      "description": "Plugin configuration exposed to the UI"
    },
    {
      "name": "icon",
      "description": "Custom own-boat icon shown on the map"
    }
  ],
  "components": {
    "schemas": {
      "SuccessResponse": {
        "type": "object",
        "required": ["statusCode", "state"],
        "properties": {
          "statusCode": {
            "type": "integer",
            "example": 200
          },
          "state": {
            "type": "string",
            "enum": ["COMPLETED"]
          }
        }
      },
      "Config": {
        "type": "object",
        "description": "Per-user UI preferences. Stored per identity (username for user tokens, clientId for device tokens, a shared anonymous bucket otherwise); GET resolves defaults, then the boat-wide baseline migrated from pre-1.5 plugin config, then the identity's own saved keys.",
        "properties": {
          "defaultBasemap": {
            "type": "string",
            "enum": ["OpenStreetMap", "Satellite", "Blank"],
            "description": "Which map layer to show on load. \"Blank\" draws no tiles (offline / local-charts-only use)."
          },
          "fleetFilterRadius": {
            "type": "integer",
            "minimum": 0,
            "description": "Radius (meters) around own vessel for displaying other vessels and historical tracks."
          },
          "enableBoatLabels": {
            "type": "boolean",
            "description": "Show other vessels' names as labels on the map (only once zoomed in enough to be legible)."
          },
          "enableOwnBoatLabel": {
            "type": "boolean",
            "description": "Show your own vessel's name as a label on the map. Independent of enableBoatLabels; the own label always wins a label collision."
          },
          "enableOwnTrack": {
            "type": "boolean",
            "description": "Draw your own vessel's historical track on the map."
          },
          "enableOtherTracks": {
            "type": "boolean",
            "description": "Draw other vessels' historical tracks on the map."
          },
          "enableRoutes": {
            "type": "boolean",
            "description": "Draw routes from the SignalK resources API on the map (read-only). Requires a resources provider plugin serving /resources/routes."
          },
          "enableChartLayers": {
            "type": "boolean",
            "description": "When local raster charts are available (via a charts plugin), enable them as map overlays by default."
          },
          "enableSeascape": {
            "type": "boolean",
            "description": "Overlay the Seascape bathymetry (depth) chart on the base map by default. Requires internet + WebGL; toggleable at runtime in the layer control."
          },
          "enableLargeControls": {
            "type": "boolean",
            "description": "Draw the map control buttons at 1.5x size for easier touch targets; off restores the original smaller sizing."
          },
          "courseVectorMinutes": {
            "type": "integer",
            "enum": [0, 5, 10, 15, 30, 60],
            "description": "Length, in minutes of travel, of the COG/SOG course vector (predictor line) drawn from each vessel's bow. 0 disables it."
          },
          "enableLookAhead": {
            "type": "boolean",
            "description": "While following your boat, bias the view toward the water ahead based on speed and course instead of centering on the boat."
          },
          "charts": {
            "type": "object",
            "additionalProperties": {
              "type": "boolean"
            },
            "description": "Per-chart show/hide choices for SignalK-provided chart overlays, keyed by chart identifier. A chart with no entry defaults to enabled. Usually updated one chart at a time via POST /ui-config/charts as the user ticks layer-control checkboxes."
          },
          "hasCustomIcon": {
            "type": "boolean",
            "readOnly": true,
            "description": "Derived (not stored): whether a custom own-boat icon has been uploaded via PUT /icon. The UI uses it to decide whether to render the custom icon and how to seed the settings preview."
          },
          "version": {
            "type": "string",
            "readOnly": true,
            "description": "Derived (not stored): the plugin version, shown in the settings footer."
          }
        }
      },
      "ErrorResponse": {
        "type": "object",
        "required": ["statusCode", "state", "message"],
        "properties": {
          "statusCode": {
            "type": "integer",
            "example": 403
          },
          "state": {
            "type": "string",
            "enum": ["FAILED"]
          },
          "message": {
            "type": "string"
          }
        }
      }
    },
    "responses": {
      "PluginError": {
        "description": "Request rejected due to a validation error.",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/ErrorResponse"
            },
            "example": {
              "statusCode": 403,
              "state": "FAILED",
              "message": "fleetFilterRadius must be a number"
            }
          }
        }
      },
      "InternalError": {
        "description": "Unexpected internal error.",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/ErrorResponse"
            },
            "example": {
              "statusCode": 500,
              "state": "FAILED",
              "message": "internal error"
            }
          }
        }
      }
    }
  },
  "paths": {
    "/ui-config": {
      "get": {
        "tags": ["config"],
        "summary": "Get the requesting identity's UI preferences",
        "description": "Returns the effective UI preferences for the requesting identity (username for user tokens, device clientId for device tokens, a shared anonymous bucket otherwise): schema defaults, overlaid with the boat-wide baseline, overlaid with the identity's own saved keys. Every preference key is always present. Requires readonly permission.",
        "responses": {
          "200": {
            "description": "Effective UI preferences returned successfully.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Config"
                },
                "example": {
                  "fleetFilterRadius": 500
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": ["config"],
        "summary": "Update the requesting identity's UI preferences",
        "description": "Persists UI-editable settings for the requesting identity. Only whitelisted properties are accepted; each is validated and merged into that identity's preference file — other identities are unaffected. Requires readwrite permission.",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Config"
              },
              "example": {
                "enableBoatLabels": false,
                "fleetFilterRadius": 750
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Configuration saved successfully.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SuccessResponse"
                }
              }
            }
          },
          "403": {
            "$ref": "#/components/responses/PluginError"
          },
          "500": {
            "$ref": "#/components/responses/InternalError"
          }
        }
      }
    },
    "/ui-config/charts": {
      "post": {
        "tags": ["config"],
        "summary": "Set one chart overlay's show/hide preference",
        "description": "Persists a single SignalK-provided chart overlay's enabled/disabled choice into the requesting identity's `charts` preference map, leaving the identity's other chart choices untouched. Charts are dynamic (only those covering the current view are listed), so the client names just the one chart it toggled instead of echoing the whole map. A chart with no entry defaults to enabled. Requires readwrite permission.",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["identifier", "enabled"],
                "properties": {
                  "identifier": {
                    "type": "string",
                    "description": "The chart's identifier from the SignalK charts catalog."
                  },
                  "enabled": {
                    "type": "boolean",
                    "description": "Whether the chart overlay should be shown on the map."
                  }
                }
              },
              "example": {
                "identifier": "Fiji_Nanuku-Passage",
                "enabled": false
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Preference saved successfully.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SuccessResponse"
                }
              }
            }
          },
          "403": {
            "$ref": "#/components/responses/PluginError"
          },
          "500": {
            "$ref": "#/components/responses/InternalError"
          }
        }
      }
    },
    "/icon": {
      "get": {
        "tags": ["icon"],
        "summary": "Get the custom own-boat icon",
        "description": "Serves the uploaded own-boat icon image. Returns 404 when no custom icon has been set. Requires readonly permission.",
        "responses": {
          "200": {
            "description": "The custom icon image.",
            "content": {
              "image/png": {
                "schema": { "type": "string", "format": "binary" }
              },
              "image/jpeg": {
                "schema": { "type": "string", "format": "binary" }
              },
              "image/gif": {
                "schema": { "type": "string", "format": "binary" }
              },
              "image/webp": {
                "schema": { "type": "string", "format": "binary" }
              }
            }
          },
          "404": {
            "description": "No custom icon is set.",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ErrorResponse" },
                "example": {
                  "statusCode": 404,
                  "state": "FAILED",
                  "message": "no custom boat icon set"
                }
              }
            }
          }
        }
      },
      "put": {
        "tags": ["icon"],
        "summary": "Upload (or replace) the custom own-boat icon",
        "description": "Stores the request body as the own-boat icon, overwriting any existing one. The body is the raw image bytes; the type is detected from the file's magic bytes (the Content-Type header is not trusted). Accepts jpg, png, gif, or webp up to 500 KB. Requires admin permission.",
        "requestBody": {
          "required": true,
          "content": {
            "image/png": { "schema": { "type": "string", "format": "binary" } },
            "image/jpeg": {
              "schema": { "type": "string", "format": "binary" }
            },
            "image/gif": { "schema": { "type": "string", "format": "binary" } },
            "image/webp": { "schema": { "type": "string", "format": "binary" } }
          }
        },
        "responses": {
          "200": {
            "description": "Icon stored successfully.",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/SuccessResponse" }
              }
            }
          },
          "400": {
            "description": "Empty upload.",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ErrorResponse" },
                "example": {
                  "statusCode": 400,
                  "state": "FAILED",
                  "message": "empty upload"
                }
              }
            }
          },
          "413": {
            "description": "Image exceeds the 500 KB size limit.",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ErrorResponse" },
                "example": {
                  "statusCode": 413,
                  "state": "FAILED",
                  "message": "image too large — max 500 KB"
                }
              }
            }
          },
          "415": {
            "description": "Body is not a recognized jpg/png/gif/webp image.",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ErrorResponse" },
                "example": {
                  "statusCode": 415,
                  "state": "FAILED",
                  "message": "unsupported image type — use jpg, png, gif, or webp"
                }
              }
            }
          },
          "500": {
            "$ref": "#/components/responses/InternalError"
          }
        }
      },
      "delete": {
        "tags": ["icon"],
        "summary": "Delete the custom own-boat icon",
        "description": "Removes any stored own-boat icon. Idempotent — succeeds whether or not an icon was set. The map reverts to the AIS ship-type icon. Requires admin permission.",
        "responses": {
          "200": {
            "description": "Icon removed (or none was set).",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/SuccessResponse" }
              }
            }
          },
          "500": {
            "$ref": "#/components/responses/InternalError"
          }
        }
      }
    }
  }
}